+++ data-raw/create_data.R
@@ -9,7 +9,7 @@ data_scripts <- list.files(
   pattern = "\\.R$",
   full.names = TRUE
 )
-
+data_scripts <- data_scripts[8]
 # Filter out helper and create_data scripts
 data_scripts <- data_scripts[
   !grepl("(helpers\\.R|create_data\\.R)", data_scripts)