+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated source file OCMQs to reflect FDA changes to FMQ
 - Renamed adaefmq to adaeocmq
 - Fixed ADCM to properly handle uncoded terms
+- Fixed ADLB to use "BEFORE TREATMENT" for ATPT instead of "00:00"
+- ADSL: Set RANDFL to "Y" for most subjects with SAFFL="Y", but also for a few with SAFFL="N" and non-NA TRT01P
 
 ## [0.0.1] - 2024-07-02